Move the memory resource header up a level
Previously this was sitting all by itself in its own directory, for no
good reason. We move it down to fit in with the rest of the
memory-related header files.
