+++ core/src/memory/allocator.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "vecmem/memory/allocator.hpp"
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 
 namespace vecmem {
     allocator::allocator(
