+++ cuda/include/vecmem/memory/cuda/host_memory_resource.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 
 namespace vecmem::cuda {
     /**
