+++ cuda/include/vecmem/utils/cuda/copy.hpp
@@ -10,7 +10,7 @@
 // Local include(s).
 #include "vecmem/containers/details/vector_buffer.hpp"
 #include "vecmem/containers/details/vector_view.hpp"
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 
 namespace vecmem::cuda {
