+++ cuda/src/memory/cuda/device_memory_resource.cpp
@@ -10,7 +10,7 @@
 #include "../../utils/cuda_wrappers.hpp"
 #include "../../utils/select_device.hpp"
 
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 #include "vecmem/memory/cuda/device_memory_resource.hpp"
 
 #include <cuda_runtime_api.h>
