+++ cuda/src/memory/cuda/host_memory_resource.cpp
@@ -8,7 +8,7 @@
 
 #include "../../utils/cuda_error_handling.hpp"
 
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 #include "vecmem/memory/cuda/host_memory_resource.hpp"
 
 #include <cuda_runtime_api.h>
