+++ sycl/include/vecmem/memory/sycl/details/memory_resource_base.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 // Local include(s).
-#include "vecmem/memory/resources/memory_resource.hpp"
+#include "vecmem/memory/memory_resource.hpp"
 #include "vecmem/utils/sycl/queue_wrapper.hpp"
 
 namespace vecmem::sycl::details {
